# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Taskr API",
...
        "contact": {
            "name": "API Support",
            "email": "support@taskr.com"
        },
        "version": "v1.0.0"
    },
    "paths": {
...
        "/v1/files": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Upload one or more files.",
                "description": "Accepts a `multipart/form-data` body with one or more files in the\r\n`files` field. Each file must be one of the allowed content types\r\nand no larger than 10 MB. The returned URL is what should be stored on\r\nresources (avatar, cover image, task attachments, etc.).",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                }
                            },
                            "encoding": {
                                "files": {
                                    "style": "form"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileResponseApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileResponseApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileResponseApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Content Too Large",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Files"
                ],
                "summary": "Delete a previously uploaded file by URL.",
                "description": "Removes the underlying object from the configured storage provider.\r\nThe URL must be one previously returned by the upload endpoint and\r\nowned by the calling user.",
                "parameters": [
                    {
                        "name": "url",
                        "in": "query",
                        "schema": {
                            "title": "url",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ObjectApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FileResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Server-generated identifier for the file record.",
                        "nullable": true
                    },
                    "url": {
                        "type": "string",
                        "description": "Publicly accessible URL where the file can be fetched. Store this on the owning resource.",
                        "nullable": true
                    },
                    "originalFilename": {
                        "type": "string",
                        "description": "The filename provided by the client at upload time.",
                        "nullable": true
                    },
                    "fileSize": {
                        "type": "integer",
                        "description": "File size in bytes.",
                        "format": "int64"
                    },
                    "fileSizeDisplay": {
                        "type": "string",
                        "description": "Human-readable file size (e.g. `12.1 KB`).",
                        "nullable": true,
                        "readOnly": true
                    },
                    "contentType": {
                        "type": "string",
                        "description": "MIME type of the uploaded file.",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "description": "Timestamp (UTC) the file was uploaded.",
                        "format": "date-time"
                    }
                },
                "additionalProperties": false,
                "description": "Response for a single uploaded file."
            },
            "FileResponseApiResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "Always `true` for success responses."
                    },
                    "message": {
                        "type": "string",
                        "description": "Short human-readable status message (e.g. `Operation successful`).",
                        "nullable": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/FileResponse"
                    },
                    "errors": {
                        "description": "Field-level errors, or `null` for success responses.",
                        "nullable": true
                    }
                },
                "additionalProperties": false,
                "description": "Standard success response envelope. Every successful API response is\r\nwrapped in this shape so the client always sees the same top-level\r\nstructure."
            },
...
            "ObjectApiResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "Always `true` for success responses."
                    },
                    "message": {
                        "type": "string",
                        "description": "Short human-readable status message (e.g. `Operation successful`).",
                        "nullable": true
                    },
                    "data": {
                        "description": "Response payload. For paginated responses this is the items list directly.",
                        "nullable": true
                    },
                    "errors": {
                        "description": "Field-level errors, or `null` for success responses.",
                        "nullable": true
                    }
                },
                "additionalProperties": false,
                "description": "Standard success response envelope. Every successful API response is\r\nwrapped in this shape so the client always sees the same top-level\r\nstructure."
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}